requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
matplotlib>=3.4.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0
Pillow>=8.3.0
opencv-python>=4.5.0

# For SegFormer model
transformers>=4.16.0
accelerate>=0.6.0

# For data handling
albumentations>=1.1.0

# Optional dependencies
jupyter>=1.0.0
seaborn>=0.11.2
